refactor(core): lift meta-card out of list framework, drop ngrx fallback
