+++ src/frontend/packages/cloud-foundry/src/shared/components/list/list-types/app/card/card-app.component.ts
@@ -10,11 +10,11 @@ import { CFAppState } from '../../../../../../../../cloud-foundry/src/cf-app-sta
 import { applicationEntityType } from '../../../../../../../../cloud-foundry/src/cf-entity-types';
 import { ApplicationStateComponent } from '../../../../../../../../core/src/shared/components/application-state/application-state.component';
 import { CardCell } from '../../../../../../../../core/src/shared/components/list/list.types';
-import { MetaCardComponent } from '../../../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-base/meta-card.component';
-import { MetaCardItemComponent } from '../../../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-item/meta-card-item.component';
-import { MetaCardKeyComponent } from '../../../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-key/meta-card-key.component';
-import { MetaCardTitleComponent } from '../../../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-title/meta-card-title.component';
-import { MetaCardValueComponent } from '../../../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-value/meta-card-value.component';
+import { MetaCardComponent } from '../../../../../../../../core/src/shared/components/meta-card/meta-card-base/meta-card.component';
+import { MetaCardItemComponent } from '../../../../../../../../core/src/shared/components/meta-card/meta-card-item/meta-card-item.component';
+import { MetaCardKeyComponent } from '../../../../../../../../core/src/shared/components/meta-card/meta-card-key/meta-card-key.component';
+import { MetaCardTitleComponent } from '../../../../../../../../core/src/shared/components/meta-card/meta-card-title/meta-card-title.component';
+import { MetaCardValueComponent } from '../../../../../../../../core/src/shared/components/meta-card/meta-card-value/meta-card-value.component';
 import { MultilineTitleComponent } from '../../../../../../../../core/src/shared/components/multiline-title/multiline-title.component';
 import { APIResource } from '../../../../../../../../store/src/types/api.types';
 import { ComponentEntityMonitorConfig, StratosStatus } from '../../../../../../../../store/src/types/shared.types';
