+++ src/frontend/packages/core/src/shared/components/meta-card/meta-card-base/meta-card.component.scss
