+++ src/frontend/packages/core/src/shared/components/meta-card/meta-card-item/meta-card-item.component.ts
