+++ src/frontend/packages/core/src/shared/components/meta-card/meta-card-key/meta-card-key.component.spec.ts
