+++ src/frontend/packages/core/src/shared/components/meta-card/meta-card-title/meta-card-title.component.scss
