+++ src/frontend/packages/core/src/shared/components/meta-card/meta-card-value/meta-card-value.component.spec.ts
