+++ src/frontend/packages/core/src/shared/shared.module.ts
@@ -37,11 +37,11 @@ import { FocusDirective } from './components/focus.directive';
 import { IntroScreenComponent } from './components/intro-screen/intro-screen.component';
 import { JsonViewerComponent } from './components/json-viewer/json-viewer.component';
 import { listCardComponents } from './components/list/list-cards/card.types';
-import { MetaCardComponent } from './components/list/list-cards/meta-card/meta-card-base/meta-card.component';
-import { MetaCardItemComponent } from './components/list/list-cards/meta-card/meta-card-item/meta-card-item.component';
-import { MetaCardKeyComponent } from './components/list/list-cards/meta-card/meta-card-key/meta-card-key.component';
-import { MetaCardTitleComponent } from './components/list/list-cards/meta-card/meta-card-title/meta-card-title.component';
-import { MetaCardValueComponent } from './components/list/list-cards/meta-card/meta-card-value/meta-card-value.component';
+import { MetaCardComponent } from './components/meta-card/meta-card-base/meta-card.component';
+import { MetaCardItemComponent } from './components/meta-card/meta-card-item/meta-card-item.component';
+import { MetaCardKeyComponent } from './components/meta-card/meta-card-key/meta-card-key.component';
+import { MetaCardTitleComponent } from './components/meta-card/meta-card-title/meta-card-title.component';
+import { MetaCardValueComponent } from './components/meta-card/meta-card-value/meta-card-value.component';
 import { EntityListViewComponent } from './components/list/list-generics/entity-list-view/entity-list-view.component';
 import { ListHostDirective } from './components/list/list-generics/helpers/list-host.directive';
 import { ListViewComponent } from './components/list/list-generics/list-view/list-view.component';
