+++ src/frontend/packages/kubernetes/src/kubernetes/workloads/list-types/helm-release-card/helm-release-card.component.ts
@@ -3,11 +3,11 @@ import {Component, Input, inject, ChangeDetectionStrategy } from '@angular/core'
 import { RouterModule } from '@angular/router';
 
 import { CardCell } from '../../../../../../core/src/shared/components/list/list.types';
-import { MetaCardComponent } from '../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-base/meta-card.component';
-import { MetaCardItemComponent } from '../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-item/meta-card-item.component';
-import { MetaCardKeyComponent } from '../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-key/meta-card-key.component';
-import { MetaCardTitleComponent } from '../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-title/meta-card-title.component';
-import { MetaCardValueComponent } from '../../../../../../core/src/shared/components/list/list-cards/meta-card/meta-card-value/meta-card-value.component';
+import { MetaCardComponent } from '../../../../../../core/src/shared/components/meta-card/meta-card-base/meta-card.component';
+import { MetaCardItemComponent } from '../../../../../../core/src/shared/components/meta-card/meta-card-item/meta-card-item.component';
+import { MetaCardKeyComponent } from '../../../../../../core/src/shared/components/meta-card/meta-card-key/meta-card-key.component';
+import { MetaCardTitleComponent } from '../../../../../../core/src/shared/components/meta-card/meta-card-title/meta-card-title.component';
+import { MetaCardValueComponent } from '../../../../../../core/src/shared/components/meta-card/meta-card-value/meta-card-value.component';
 import { MultilineTitleComponent } from '../../../../../../core/src/shared/components/multiline-title/multiline-title.component';
 import { TableCellEndpointNameComponent } from '../../../../../../core/src/shared/components/list/list-types/endpoint/table-cell-endpoint-name/table-cell-endpoint-name.component';
 import { CustomIconComponent } from '../../../../../../core/src/shared/components/custom-material/custom-material.component';
